refactor: remove enc_publish functions and create separate encoding functions instead
