+++ library/c/api_relay.go
@@ -31,32 +31,6 @@ func waku_relay_publish(messageJSON *C.char, topic *C.char, ms C.int, onOkCb C.W
 	}, onOkCb, onErrCb)
 }
 
-// Publish a message encrypted with a secp256k1 public key using waku relay and returns the message ID. Use NULL for topic to use the default pubsub topic.
-// publicKey must be a hex string prefixed with "0x" containing a valid secp256k1 public key.
-// optionalSigningKey is an optional hex string prefixed with "0x" containing a valid secp256k1 private key for signing the message. Use NULL otherwise
-// If ms is greater than 0, the broadcast of the message must happen before the timeout
-// (in milliseconds) is reached, or an error will be returned.
-//
-//export waku_relay_publish_enc_asymmetric
-func waku_relay_publish_enc_asymmetric(messageJSON *C.char, topic *C.char, publicKey *C.char, optionalSigningKey *C.char, ms C.int, onOkCb C.WakuCallBack, onErrCb C.WakuCallBack) C.int {
-	return singleFnExec(func() (string, error) {
-		return library.RelayPublishEncodeAsymmetric(C.GoString(messageJSON), C.GoString(topic), C.GoString(publicKey), C.GoString(optionalSigningKey), int(ms))
-	}, onOkCb, onErrCb)
-}
-
-// Publish a message encrypted with a 32 bytes symmetric key using waku relay and returns the message ID. Use NULL for topic to use the default pubsub topic.
-// symmetricKey must be a hex string prefixed with "0x" containing a 32 bytes symmetric key
-// optionalSigningKey is an optional hex string prefixed with "0x" containing a valid secp256k1 private key for signing the message. Use NULL otherwise
-// If ms is greater than 0, the broadcast of the message must happen before the timeout
-// (in milliseconds) is reached, or an error will be returned.
-//
-//export waku_relay_publish_enc_symmetric
-func waku_relay_publish_enc_symmetric(messageJSON *C.char, topic *C.char, symmetricKey *C.char, optionalSigningKey *C.char, ms C.int, onOkCb C.WakuCallBack, onErrCb C.WakuCallBack) C.int {
-	return singleFnExec(func() (string, error) {
-		return library.RelayPublishEncodeSymmetric(C.GoString(messageJSON), C.GoString(topic), C.GoString(symmetricKey), C.GoString(optionalSigningKey), int(ms))
-	}, onOkCb, onErrCb)
-}
-
 // Subscribe to a WakuRelay topic. Set the topic to NULL to subscribe
 // to the default topic. Returns a json response. When a message is received,
 // a "message" event is emitted containing the message and pubsub topic in which
