+++ library/node.go
@@ -3,11 +3,9 @@ package library
 import (
 	"context"
 	"crypto/ecdsa"
-	"crypto/elliptic"
 	"crypto/rand"
 	"database/sql"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
@@ -17,7 +15,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/crypto/secp256k1"
 	"github.com/ethereum/go-ethereum/p2p/enode"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/libp2p/go-libp2p/core/peer"
@@ -26,7 +23,6 @@ import (
 	"github.com/waku-org/go-waku/waku/persistence"
 	dbutils "github.com/waku-org/go-waku/waku/persistence/utils"
 	"github.com/waku-org/go-waku/waku/v2/node"
-	"github.com/waku-org/go-waku/waku/v2/payload"
 	"github.com/waku-org/go-waku/waku/v2/peerstore"
 	"github.com/waku-org/go-waku/waku/v2/protocol"
 	"github.com/waku-org/go-waku/waku/v2/protocol/pb"
@@ -384,120 +380,4 @@ func Peers() (string, error) {
 	}
 
 	return marshalJSON(peers)
-}
-
-func unmarshalPubkey(pub []byte) (ecdsa.PublicKey, error) {
-	x, y := elliptic.Unmarshal(secp256k1.S256(), pub)
-	if x == nil {
-		return ecdsa.PublicKey{}, errors.New("invalid public key")
-	}
-	return ecdsa.PublicKey{Curve: secp256k1.S256(), X: x, Y: y}, nil
-}
-
-func extractPubKeyAndSignature(payload *payload.DecodedPayload) (pubkey string, signature string) {
-	pkBytes := crypto.FromECDSAPub(payload.PubKey)
-	if len(pkBytes) != 0 {
-		pubkey = hexutil.Encode(pkBytes)
-	}
-
-	if len(payload.Signature) != 0 {
-		signature = hexutil.Encode(payload.Signature)
-	}
-
-	return
-}
-
-// DecodeSymmetric decodes a waku message using a 32 bytes symmetric key. The key must be a hex encoded string with "0x" prefix
-func DecodeSymmetric(messageJSON string, symmetricKey string) (string, error) {
-	var msg pb.WakuMessage
-	err := json.Unmarshal([]byte(messageJSON), &msg)
-	if err != nil {
-		return "", err
-	}
-
-	if msg.Version == 0 {
-		return marshalJSON(msg.Payload)
-	} else if msg.Version > 1 {
-		return "", errors.New("unsupported wakumessage version")
-	}
-
-	keyInfo := &payload.KeyInfo{
-		Kind: payload.Symmetric,
-	}
-
-	keyInfo.SymKey, err = utils.DecodeHexString(symmetricKey)
-	if err != nil {
-		return "", err
-	}
-
-	payload, err := payload.DecodePayload(&msg, keyInfo)
-	if err != nil {
-		return "", err
-	}
-
-	pubkey, signature := extractPubKeyAndSignature(payload)
-
-	response := struct {
-		PubKey    string `json:"pubkey,omitempty"`
-		Signature string `json:"signature,omitempty"`
-		Data      []byte `json:"data"`
-		Padding   []byte `json:"padding"`
-	}{
-		PubKey:    pubkey,
-		Signature: signature,
-		Data:      payload.Data,
-		Padding:   payload.Padding,
-	}
-
-	return marshalJSON(response)
-}
-
-// DecodeAsymmetric decodes a waku message using a secp256k1 private key. The key must be a hex encoded string with "0x" prefix
-func DecodeAsymmetric(messageJSON string, privateKey string) (string, error) {
-	var msg pb.WakuMessage
-	err := json.Unmarshal([]byte(messageJSON), &msg)
-	if err != nil {
-		return "", err
-	}
-
-	if msg.Version == 0 {
-		return marshalJSON(msg.Payload)
-	} else if msg.Version > 1 {
-		return "", errors.New("unsupported wakumessage version")
-	}
-
-	keyInfo := &payload.KeyInfo{
-		Kind: payload.Asymmetric,
-	}
-
-	keyBytes, err := utils.DecodeHexString(privateKey)
-	if err != nil {
-		return "", err
-	}
-
-	keyInfo.PrivKey, err = crypto.ToECDSA(keyBytes)
-	if err != nil {
-		return "", err
-	}
-
-	payload, err := payload.DecodePayload(&msg, keyInfo)
-	if err != nil {
-		return "", err
-	}
-
-	pubkey, signature := extractPubKeyAndSignature(payload)
-
-	response := struct {
-		PubKey    string `json:"pubkey,omitempty"`
-		Signature string `json:"signature,omitempty"`
-		Data      []byte `json:"data"`
-		Padding   []byte `json:"padding"`
-	}{
-		PubKey:    pubkey,
-		Signature: signature,
-		Data:      payload.Data,
-		Padding:   payload.Padding,
-	}
-
-	return marshalJSON(response)
 }
